requirements.txt
openai
click
pyyaml
python-dotenv
pydub
sounddevice
numpy
faster-whisper
